importer/fs: dedup walkDir_walker and inline namecache
