+++ snapshot/importer/fs/walkdir.go
@@ -21,7 +21,6 @@ package fs
 
 import (
 	"fmt"
-	"io/fs"
 	"os"
 	"os/user"
 	"path/filepath"
@@ -33,15 +32,8 @@ import (
 	"github.com/pkg/xattr"
 )
 
-type namecache struct {
-	uidToName map[uint64]string
-	gidToName map[uint64]string
-
-	mu sync.RWMutex
-}
-
 // Worker pool to handle file scanning in parallel
-func walkDir_worker(jobs <-chan string, results chan<- *importer.ScanResult, wg *sync.WaitGroup, namecache *namecache) {
+func (f *FSImporter) walkDir_worker(jobs <-chan string, results chan<- *importer.ScanResult, wg *sync.WaitGroup) {
 	defer wg.Done()
 
 	for path := range jobs {
@@ -59,35 +51,35 @@ func walkDir_worker(jobs <-chan string, results chan<- *importer.ScanResult, wg 
 
 		fileinfo := objects.FileInfoFromStat(info)
 
-		namecache.mu.RLock()
-		if uname, ok := namecache.uidToName[fileinfo.Uid()]; !ok {
+		f.mu.RLock()
+		if uname, ok := f.uidToName[fileinfo.Uid()]; !ok {
 			if u, err := user.LookupId(fmt.Sprintf("%d", fileinfo.Uid())); err == nil {
 				fileinfo.Lusername = u.Username
 
-				namecache.mu.RUnlock()
-				namecache.mu.Lock()
-				namecache.uidToName[fileinfo.Uid()] = u.Username
-				namecache.mu.Unlock()
-				namecache.mu.RLock()
+				f.mu.RUnlock()
+				f.mu.Lock()
+				f.uidToName[fileinfo.Uid()] = u.Username
+				f.mu.Unlock()
+				f.mu.RLock()
 			}
 		} else {
 			fileinfo.Lusername = uname
 		}
 
-		if gname, ok := namecache.gidToName[fileinfo.Gid()]; !ok {
+		if gname, ok := f.gidToName[fileinfo.Gid()]; !ok {
 			if g, err := user.LookupGroupId(fmt.Sprintf("%d", fileinfo.Gid())); err == nil {
 				fileinfo.Lgroupname = g.Name
 
-				namecache.mu.RUnlock()
-				namecache.mu.Lock()
-				namecache.gidToName[fileinfo.Gid()] = g.Name
-				namecache.mu.Unlock()
-				namecache.mu.RLock()
+				f.mu.RUnlock()
+				f.mu.Lock()
+				f.gidToName[fileinfo.Gid()] = g.Name
+				f.mu.Unlock()
+				f.mu.RLock()
 			}
 		} else {
 			fileinfo.Lgroupname = gname
 		}
-		namecache.mu.RUnlock()
+		f.mu.RUnlock()
 
 		var originFile string
 		if fileinfo.Mode()&os.ModeSymlink != 0 {
@@ -121,47 +113,4 @@ func walkDir_addPrefixDirectories(rootDir string, jobs chan<- string, results ch
 
 		jobs <- path
 	}
-}
-
-func (f *FSImporter) walkDir_walker(results chan<- *importer.ScanResult, rootDir string, numWorkers int) {
-	namecache := &namecache{
-		uidToName: make(map[uint64]string),
-		gidToName: make(map[uint64]string),
-	}
-
-	real, err := f.realpathFollow(rootDir)
-	if err != nil {
-		results <- importer.NewScanError(rootDir, err)
-		return
-	}
-
-	jobs := make(chan string, 1000) // Buffered channel to feed paths to workers
-	var wg sync.WaitGroup
-	for range numWorkers {
-		wg.Add(1)
-		go walkDir_worker(jobs, results, &wg, namecache)
-	}
-
-	// Add prefix directories first
-	walkDir_addPrefixDirectories(real, jobs, results)
-	if real != rootDir {
-		jobs <- rootDir
-		walkDir_addPrefixDirectories(rootDir, jobs, results)
-	}
-
-	err = filepath.WalkDir(real, func(path string, d fs.DirEntry, err error) error {
-		if err != nil {
-			results <- importer.NewScanError(path, err)
-			return nil
-		}
-		jobs <- path
-		return nil
-	})
-	if err != nil {
-		results <- importer.NewScanError(real, err)
-	}
-
-	close(jobs)
-	wg.Wait()
-	close(results)
 }
