+++ snapshot/importer/fs/walkdir_windows.go
@@ -44,7 +44,7 @@ func toUnixPath(pathname string) string {
 }
 
 // Worker pool to handle file scanning in parallel
-func walkDir_worker(jobs <-chan string, results chan<- *importer.ScanResult, wg *sync.WaitGroup) {
+func (f *FSImporter) walkDir_worker(jobs <-chan string, results chan<- *importer.ScanResult, wg *sync.WaitGroup) {
 	defer wg.Done()
 
 	for pathname := range jobs {
